Merge most recent into dev-bergen bouvet

+++ property/inc/class.soworkorder.inc.php
@@ -1460,10 +1460,10 @@
 
 			if (isset($GLOBALS['phpgw_info']['user']['preferences']['common']['currency']))
 			{
-				$workorder['contract_sum'] = str_ireplace($GLOBALS['phpgw_info']['user']['preferences']['common']['currency'], '', $workorder['contract_sum']);
+				$workorder['contract_sum'] = str_ireplace($GLOBALS['phpgw_info']['user']['preferences']['common']['currency'], '', (float)$workorder['contract_sum']);
 			}
 
-			$workorder['contract_sum'] = str_replace(array(' ', ','), array('', '.'), $workorder['contract_sum']);
+			$workorder['contract_sum'] = (float)str_replace(array(' ', ','), array('', '.'), $workorder['contract_sum']);
 
 			if (abs((int)$workorder['contract_sum']) > 0)
 			{
@@ -1627,13 +1627,13 @@
 			{
 				$this->db->query("SELECT sum(budget) AS budget FROM fm_workorder_budget WHERE order_id = '{$workorder['id']}'", __LINE__, __FILE__);
 				$this->db->next_record();
-				$old_budget = $this->db->f('budget');
+				$old_budget = (float)$this->db->f('budget');
 
 				$this->delete_period_from_budget($workorder['id'], $workorder['delete_b_period']);
 
 				$this->db->query("SELECT sum(budget) AS budget FROM fm_workorder_budget WHERE order_id = '{$workorder['id']}'", __LINE__, __FILE__);
 				$this->db->next_record();
-				$new_budget = $this->db->f('budget');
+				$new_budget = (float)$this->db->f('budget');
 
 				$historylog->add('B', $workorder['id'], $new_budget, $old_budget);
 			}
@@ -1643,13 +1643,13 @@
 			{
 				$this->db->query("SELECT sum(budget) AS budget FROM fm_workorder_budget WHERE order_id = '{$workorder['id']}'", __LINE__, __FILE__);
 				$this->db->next_record();
-				$old_budget = $this->db->f('budget');
+				$old_budget = (float)$this->db->f('budget');
 
-				$this->_update_order_budget($workorder['id'], $workorder['budget_year'], $periodization_id, $workorder['budget'], $workorder['contract_sum'], $combined_cost);
+				$this->_update_order_budget($workorder['id'], $workorder['budget_year'], $periodization_id, (float)$workorder['budget'], (float)$workorder['contract_sum'], (float)$combined_cost);
 
 				$this->db->query("SELECT sum(budget) AS budget FROM fm_workorder_budget WHERE order_id = '{$workorder['id']}'", __LINE__, __FILE__);
 				$this->db->next_record();
-				$new_budget = $this->db->f('budget');
+				$new_budget = (float)$this->db->f('budget');
 
 				$historylog->add('B', $workorder['id'], $new_budget, $old_budget);
 			}
@@ -2954,9 +2954,9 @@
 			{
 				$start_date			 = $this->db->f('start_date');
 				$periodization_id	 = (int)$this->db->f('periodization_id');
-				$budget				 = $this->db->f('budget');
-				$contract_sum		 = $this->db->f('contract_sum');
-				$combined_cost		 = $this->db->f('combined_cost');
+				$budget				 = (float)$this->db->f('budget');
+				$contract_sum		 = (float)$this->db->f('contract_sum');
+				$combined_cost		 = (float)$this->db->f('combined_cost');
 				$this->_update_order_budget($order_id, date('Y', $start_date), $periodization_id, $budget, $contract_sum, $combined_cost);
 			}
 		}
@@ -3050,10 +3050,10 @@
 				{
 					$periodization_outline[] = array
 						(
-						'month'		 => $this->db->f('month'),
-						'value'		 => $this->db->f('value'),
-						'dividend'	 => $this->db->f('dividend'),
-						'divisor'	 => $this->db->f('divisor')
+						'month'		 => (int)$this->db->f('month'),
+						'value'		 => (float)$this->db->f('value'),
+						'dividend'	 => (int)$this->db->f('dividend'),
+						'divisor'	 => (int)$this->db->f('divisor')
 					);
 				}
 			}
